requirements.txt
paramiko==4.0.0
mysql-connector-python==9.6.0
bcrypt==5.0.0
cryptography==46.0.7
pynacl==1.6.2
cffi==2.0.0
pycparser==3.0
invoke==3.0.3
